backend/requirements.txt
flask==2.0.3
python-dotenv==0.19.2
werkzeug==2.0.3
